working sw volume change
